requirements.txt
pandas==2.2.3
langchain==0.3.18
langchain-community==0.3.17
openai==1.61.1
langchain-core~=0.3.34
python-dotenv~=1.0.1
transformers~=4.48.3
tiktoken~=0.8.0
numpy~=1.26.4
matplotlib~=3.9.4
wordcloud~=1.9.4
streamlit~=1.42.0
scikit-learn~=-1.6.1
tensorflow~=2.18.0
keras~=3.8.0
faiss_cpu~=1.10.0
seaborn~=0.13.2
nltk~=3.9.1
